Factor out schema hash
